Make installation of software-properties optional
To simplify configuration of package installs on
Ubuntu, I frequently do the following:

    Apt::Source<||> -> Package<||>

To ensure that apt source configuration occurrs before
package installation. Having the ppas install software
properties breaks this pattern because it means that
the apt-get update must be run both before and after
software properties resulting in a dependency cycle.

For my purposes, the value of being able to stage
apt operations before packages is significannt enough
that I just ensure that software properties in installed
as a part of my base image.

According to the following issue, this looks like it has
been resolved as of 14.04 (that this library is installed
by default)

  https://bugs.launchpad.net/ubuntu/+source/ubuntu-meta/+bug/439566