Allow URI escaping in any user/pass in the URL
One of my providers distributes packages from behind a password protected repository. And email addresses have become very common usernames, and the '@' in a username won't work since the URL decoder will detect that as the end of the user:pass part. This trivial patch allows both the username and the password to be URI encoded.

I'm pretty sure that doing the URI un-escaping after splitting 'parsed_value' is the right thing, in case either includes a colon ':' character.