key: Don't include dirmngr unless necessary
We should not depend on dirmngr unless the apt::key resource does not
specify a content or a source parameter. When either content or source
is set we either have a local file or will fetch one over http.

This also fixes a subtle bug related to Raspbian where dirmngr isn't
available which renders apt::key unusable even when the apt::key
or apt::source entry specifies a content/source parameter for the
key.