(maint) Update apt custom facts
This commit updates apt custom facts because the evaluation
of custom facts code has changed between facter 2 and facter 4.
In facter 2, when a custom fact is loaded, the code in `add` block
was not executed.
In facter 4, when a custom fact is loaded, the code outside `setcode`
block is executed(same as Facter 3).