(maint) Fix resource ordering when apt-transport-https is needed
On older Debian / Ubuntu releases, the apt-transport-https package is
required to download from an https:// source.  This package dependency
is ensured by ensure_packages() but no ordering is set, possibly
resulting of the resource being applied in a non-working order:

1. Setup a https:// source
2. Trigger apt update (this will fail)
3. Install apt-transport-https

Add a dependency to ensure apt-transport-https is installed when the
repositories are updated.